pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "cython", "numpy"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-quant-platform"
version = "0.1.0"
description = "AI-Powered Quantitative Research Platform - Cython/C++ extensions: LOB and microstructure generator"
dependencies = [
    "numpy>=1.24.0",
    "optuna>=3.0.0,<4.0.0",
    "pandas>=2.0.0",
    "pydantic",
    "pyyaml",
    "stable-baselines3>=2.0.0,<3.0.0",
]

[project.scripts]
no-trade-mask = "apply_no_trade_mask:main"

[project.optional-dependencies]
extra = [
    "numpy>=1.24.0",
    "optuna>=3.0.0,<4.0.0",
    "pandas>=2.0.0",
    "pyarrow>=12.0.0",
    "requests>=2.31.0",
    "investpy==1.0.8",
    "joblib>=1.3.0",
    "torch>=2.0.0 ; platform_system!='Windows'",
    "stable-baselines3>=2.0.0,<3.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "asyncio: marks tests as async (using pytest-asyncio)",
    "ppo_bugs: marks tests for PPO bug fixes",
    "integration: marks tests as integration tests",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "requires_cython: marks tests that require compiled Cython modules",
]
asyncio_mode = "auto"
